backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
Pillow==9.3.0
django-debug-toolbar==3.8.1
djoser
psycopg2-binary==2.9.3
# django-link-shortener
python-dotenv==1.0.1 
django-filter==23.1